backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-pptx==1.0.2
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
lxml==5.3.0
Pillow==10.4.0
reportlab>=4.0.0
python-dotenv==1.0.0
